Refactored DeleteAllBorrowersServiceImpl

+++ v1/src/main/java/com/api/v1/borrower/services/DeleteAllBorrowersServiceImpl.java
@@ -1,10 +1,12 @@
 package com.api.v1.borrower.services;
 
+import com.api.v1.exceptions.EmptyFluxException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.api.v1.borrower.domain.BorrowerRepository;
 
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 @Service
@@ -15,7 +17,13 @@ class DeleteAllBorrowersServiceImpl implements DeleteAllBorrowersService {
 
     @Override
     public Mono<Void> deleteAll() {
-        return repository.deleteAll();
+        return repository
+                .findAll()
+                .hasElements()
+                .flatMap(exists -> {
+                    if (!exists) return Mono.error(EmptyFluxException::new);
+                    return repository.deleteAll();
+                });
     }
     
 }
